rviz: progress on #5058, should use 3D position of mouse click to make dragging better.  DO NOT RELEASE yet, this work has introduced some bugs into the interactive marker drag behaviors.
git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/visualization/trunk@37825 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/rviz/displays_panel.cpp
@@ -142,7 +142,7 @@ void ManageDisplaysDialog::onRemove(wxCommandEvent& event)
 
   manager_->removeDisplay(displays_[sel]);
   listbox_->Delete(sel);
-  if (sel < listbox_->GetCount())
+  if (sel < (int) listbox_->GetCount())
   {
     listbox_->SetSelection(sel);
   }
@@ -173,7 +173,7 @@ void ManageDisplaysDialog::onMoveUp( wxCommandEvent& event )
 void ManageDisplaysDialog::onMoveDown( wxCommandEvent& event )
 {
   int sel = listbox_->GetSelection();
-  if (sel >= 0 && sel < listbox_->GetCount() - 1)
+  if (sel >= 0 && sel < (int) listbox_->GetCount() - 1)
   {
     std::swap(displays_[sel], displays_[sel + 1]);
     listbox_->Insert(listbox_->GetString(sel + 1), sel);
